Fix bug in WordPress_Sniffs_VIP_DirectDatabaseQuerySniff
The `WordPress.VIP.DirectDatabaseQuery` sniff allows for adding custom database query caching functions to the predefined lists of known WP core database query caching functions.

The known WP core database query caching functions are contained in static variables in the `WordPress_Sniff` base class and are currently only used by the `WordPress.VIP.DirectDatabaseQuery` sniff.

Up to now, if custom functions were passed to this sniff, it would merge them with the lists of WP core functions and then **overwrite** the values of the static properties in the `WordPress_Sniff` base class.

While no other sniff is using these properties, this is not much of a problem (yet), but as soon as another sniff would start using these properties, this behaviour would have the unintended side-effect of affecting the other sniff as well. Even worse, whether this side-effect would come into play might be unpredictable as it would depend on the running order of the sniffs which can change without notice.

To fix this, I have introduced three new private properties to the `WordPress.VIP.DirectDatabaseQuery` sniff which will hold the merged versions of the function lists.

I've also moved the setting of these properties to a dedicated function.