:rocket: Temporary solution for build failure.
As PHPCS upstream is taking its time to decide whether or not to make the test suite compatible with PHPUnit 6 or not, we might as well get the builds passing again in the mean time, so we can get the 0.11.0 release out the door.

This is intended (fingers crossed) as a temporary solution to be reverted once the testsuite for PHPCS is cross-version compatible with PHPUnit.

I've chosen to *only* download PHPUnit when we need a different version than the one installed by default in the Travis images to keep the build time down.