Fix persistent bug in VariableRestrictions abstract
While unit testing something else I ran into this bug.

Basically, the sniff allows for adding several arrays with things to be sniffed for, however when examining a `T_DOUBLE_QUOTED_STRING` or - when I add this (waiting in the backwings) - a `T_HEREDOC`, only the first set of patterns is taken into account instead of all possible patterns.

This in effect meant that the sniff suffered from false negatives.

The original intend seems quite clear what with the $patterns being merged in every part of the control structure. De-coupling the if statements solves this.

Additionally, within the last `if` structure, `$token['bracket_closer']` is referenced, while this array key will not be available when a `T_DOUBLE_QUOTED_STRING` or `T_HEREDOC` is encountered.
The `$var` setting for this is already corrected further along in the code, so making this part conditional fixes the `Undefined index` notice being thrown.