ControlStructureSpacing sniff: Sync in more changes from upstream for the "blank line after" check
The upstream sniff is *not* properly unit tested with a fixed file and in part checks for something different and therefore uses different logic.

This commit adds unit tests for our version of the sniff, syncs in a number of good changes from upstream and fixes a number of bugs related to this particular check & fixer.