ArrayDeclarationSpacing sniff: Decouple the sniff from the upstream sniff
This sniff was coupled to the upstream `Squiz_Sniffs_Arrays_ArrayDeclarationSniff` sniff to benefit from the process logic which routed the processing to the single-line/multi-line methods.

However, with the addition of the `get_function_call_parameters()` utility method which handles splitting an array into individual array items + the addition of the new `find_array_open_close()` utility method to the `WordPress_Sniff`, we don't need the upstream logic anymore and it makes more sense to decouple the sniff and let it extend the `WordPress_Sniff` instead.

Includes leveraging the properties in the `WordPress_Sniff`.