Improve detection of condition beginning in Yoda conditions sniff
This sniff is supposed to loop over the left-hand side of the condition,
checking for variables. If the left-hand side of the condition is not a
variable, then it should not give an error about it.

However, there were some issues with how we were detecting the beginning
of the left-hand side of the condition. Take the following example code:

```php
$sample = false !== strpos( $link, '%pagename%' ); // Ok.
```

This should not cause an issue, because the left-hand side of the
condition is just `false`. However, we weren't detecting the beginning
of the condition correctly; we'd end up checking right through the `=`
assignment operator and seeing the `$sample` variable. As a result, an
error would be given for this snippet.

To correct this, we just need to take assignment operators into account
when detecting the beginning of the condition.

With this correction, we can also simplify some other code that was
checking for a function call (or certain other tokens) within the
condition, that would indicate it didn't need to be checked:

- The `T_RETURN` token is actually an indicator of the beginning of the
condition, so it should be moved to that check instead.
- The `T_CONSTANT_ENCAPSED_STRING` token is no longer needed to be checked
at all, because it was added to circumvent the bug that is now fixed by
better detection of the beginning of the condition.
- The `T_OPEN_PARENTHESIS` really shouldn't be needed, because the closing
parenthesis will be detected first anyway (we are looping backward
here).

The `T_CLOSE_PARENTHESIS` is the only one that needs to stay, because of
cases like this:

```php
$sample = foo( $var ) !== strpos( $link, '%pagename%' ); // Ok.
```

Without the check for the closing parenthesis token, the `$var`
parameter in the call to the `foo()` function would cause the condition
to be flagged as needing to be a Yoda condition.

Fixes #1036