Sync the WP global var lists
The WP base `Sniff` class contains a list of WP global variables which is used by two sniffs.
The `ValidVariableName` sniff contains a secondary list of globals which don't the WP variable name guidelines.

I've done a quick compare and found some differences between the lists which needed to be fixed.

It would be nice to build up the mixed_case_vars list dynamically, but unfortunately, as the `ValidVariableName` sniff extends an upstream sniff rather than the WP `Sniff`, this is currently not an option.