Array properties: always be lenient about spaces
While PHPCS does not allow for additional spaces in custom properties which are passed through a ruleset, WPCS was already lenient about this when the property was (incorrectly) passed as a string.

This fix will also make WPCS lenient about this when the property is passed as an array.

Inspired by upstream issue squizlabs/PHP_CodeSniffer 1688

So for WPCS, all of the below four examples will now work properly as expected:
```xml
<property name="customAutoEscapedFunctions" value="foo, bar" />
<property name="customAutoEscapedFunctions" value="foo,bar" />
<property name="customAutoEscapedFunctions" value="foo, bar" type="array" />
<property name="customAutoEscapedFunctions" value="foo,bar" type="array" />
```