CronInterval: move sniff from VIP category to WP category
The original sniff class still exists, but extends the renamed sniff now.
In addition, a deprecation warning will be thrown if this sniff is included directly by a custom ruleset.
A deprecation warning will also be thrown if the sniff is included directly just to set custom properties.

Note: if the sniff was explicitly **excluded** via a custom ruleset, no deprecation warning will be thrown and the exclusion will have no effect. The custom ruleset will need to be updated to use the new sniff name for the exclusion to be reinstated.