WhiteSpace/ControlStructureSpacing: prevent fixer removing return type declarations
While this will be solved better in the future by a dedicated sniff to check function declarations, in the mean time, return type declarations should not be removed by the fixer in this sniff.

And as the placement of the opening brace for function declarations is also checked for by the `Generic.Functions.OpeningFunctionBraceKernighanRitchie` sniff, let's ignore this error for function declarations in this sniff and leave the fixing up to the upstream sniff.

Fixes 1202