ArrayIndentation: fix bug when dealing with multi-line string
If a multi-line string is broken up across several lines, the indentation of subsequent lines should not be fixed, nor should any warnings/errors be thrown.

This worked perfectly well as long as the closing quote of the multi-line string was not on a line of its own.

Now the sniff will deal correctly with that edge-case as well.

Includes unit test.