EscapeOutput: report correctly on first param of `trigger_error()`
The PHPCS `findEndOfStatement()` method returns the end token òr the last non-whitespace token if the end token is a "bracket", like a closing parenthesis.

In effect, this meant that the last token in a `trigger_error()` function call was never examined and that the sniff would underreport.

Fixed now. Includes unit tests.