Fix PHP 7.3 compatibility
As of PHP 7.3, the `#` character will also be escaped by `preg_quote()`. This was causing problems for three sniffs which used the `AbstractFunctionRestrictionsSniff::prepare_name_for_regex()` method and was the reason that the unit tests were failing in the [Travis PHP 7.3 builds](https://travis-ci.org/WordPress-Coding-Standards/WordPress-Coding-Standards/builds/389243319#ember12516).

https://github.com/php/php-src/blob/600cb4b46755a9170eaa32606528a6908bc98975/UPGRADING#L160