PrefixAllGlobals: add recording of metrics
When a project which wasn't using WPCS, starts to use it or when an project upgrades to WPCS > 0.12.0, one of the things the project most likely will run into is the prefixing of all globals.

When no prefixes are passed, the sniff will not throw any errors.

However, if the project does want to comply and wants to add a prefix configuration, it would be useful to get some insight into which prefixes are currently in use in a project and how often each is used to determine what prefixes to allow and/or what prefixes to consolidate.

PHPCS offers a metrics mechanism which can be used for this, the results of which can be viewed using the `--report-info` argument.

This PR takes advantage of that metrics report to:
1. Report on the frequency each of the allowed prefixes is used
2. Report on potential additional prefixes.

I imagine that if this PR is accepted, it would be useful to add a wiki page about how to get access to this information and how to use it.

Notes about the implementation:
* For the "potential prefixes" a relatively simple formula is used:
    - Strip off any leading underscores, namespace separators and variable `$` signs of the name as found
    - Grab the first part of the name, before the first underscore or next uppercase character.
        This matches `Pascal` in `PascalCase`, `camel` in `camelCase`, `snake` in `snake_case`.
* Potential prefixes are passed to the metrics mechanism in lowercase as (nearly all) the prefix checks are done in a case-insensitive manner.
* Potential prefixes are only recorded for those instances where an error/warning would be reported by the sniff. If a whitelist comment, a PHPCS ignore annotation or a ruleset `exclude-pattern`, prevents the error from being recorded, no "potential prefix" metric will be recorded.
* The PR changes the method signature of two methods in the sniff, but as both are `private` methods, this is not a BC-break.

- [ ] Add a wiki page about how to go about prefixing all globals in legacy projects.